webserver/pyproject.toml
[tool.poetry]
name = "math3d-webserver"
version = "0.1.0"
description = ""
authors = ["Chris Chudzicki"]
packages = [{include = "main"}]

[tool.poetry.dependencies]
python = "3.11.6"
Django = "^4.1.6"
djangorestframework = "^3.14.0"
dj-database-url = "^1.2.0"
psycopg2 = "^2.9.5"
tqdm = "^4.64.1"
jtd = "^0.1.1"
pyyaml = "^6.0"
gunicorn = "^20.1.0"
django-heroku = "^0.3.1"
django-cors-headers = "^4.1.0"
django-environ = "^0.10.0"
faker = "^19.12.0"
djoser = "^2.2.0"
drf-spectacular = "^0.26.5"
django-ses = "^3.5.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.0.0"
django-stubs = "^1.15.0"
types-tqdm = "^4.64.7.15"
types-psycopg2 = "^2.9.21.7"
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^1.9.1"}
pytest = "^7.2.2"
django = "^4.1.7"
pytest-django = "^4.5.2"
types-dj-database-url = "^1.2.0.1"
faker = "^19.12.0"
lxml = "^4.9.3"
cssselect = "^1.2.0"
lxml-stubs = "^0.4.0"
factory-boy = "^3.3.0"
types-factory-boy = "^0.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[[tool.mypy.overrides]]
module = "djoser.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "django_ses.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "templated_mail.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "environ.*"
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "main.settings"

[tool.pytest.ini_options]
testpaths = []
